Refactoring: Moved EventNames to util package
